refactor: Lines객체 삭제
